requirements.txt
customtkinter>=5.2.0
pillow>=10.0.0
requests>=2.31.0
yt-dlp>=2025.12.0
pyinstaller
watchdog
tkinterweb>=3.15.0
python-vlc
# ffmpeg se instala en el sistema (apt/brew/choco), no en pip.